% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{execute}
\alias{execute}
\title{Execute the TAXIS Phenotype Evaluation Study Package}
\usage{
execute(
  connectionDetails,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTable = "taxis_pheno_eval",
  workDatabaseSchema = cohortDatabaseSchema,
  outputFolder = "output",
  databaseId = "My_CDM",
  runCohortGeneration = TRUE,
  runCohortOverlap = TRUE,
  runDiagnostics = TRUE,
  runPheValuator = TRUE,
  minCellCount = 5
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function.}

\item{cdmDatabaseSchema}{The name of the database schema that contains the OMOP CDM
instance. Requires read permissions to this database.}

\item{cohortDatabaseSchema}{The name of the database schema where the user has write
permissions to create and write to the cohort table.}

\item{cohortTable}{The name of the cohort table to create and populate. Default is 'taxis_pheno_eval'.}

\item{workDatabaseSchema}{The name of the database schema where temporary tables can be created. Default is cohortDatabaseSchema.}

\item{outputFolder}{The path to the folder where output CSVs and ZIP archives will be saved.}

\item{databaseId}{A short identifier for the database (e.g. 'Site_EHR', 'Site_Claims').}

\item{runCohortGeneration}{Logical: should cohorts be instantiated on the CDM? Default is TRUE.}

\item{runCohortOverlap}{Logical: should pairwise overlap and Jaccard similarity be computed? Default is TRUE.}

\item{runDiagnostics}{Logical: should CohortDiagnostics be executed? Default is TRUE.}

\item{runPheValuator}{Logical: should PheValuator evaluation be performed? Default is TRUE.}

\item{minCellCount}{Minimum count threshold for cell suppression (default is 5). Counts < minCellCount are masked.}
}
\description{
Execute the TAXIS Phenotype Evaluation Study Package
}
\details{
Orchestrates cohort creation, overlap calculation, CohortDiagnostics, and PheValuator
evaluation across federated OMOP CDM partner databases.
}
